SCRUM-94: make only dfgArcs selectable
