+++ src/app/components/drawing-area/components/drawing-arc.component.ts
@@ -45,17 +45,6 @@ import { PetriNetManagementService } from 'src/app/services/petri-net-management
 
         <svg:path
             [attr.d]="pathForLine(arc)"
-            [attr.stroke-width]="10"
-            [attr.fill]="'none'"
-            fill="transparent"
-            pointer-events="all"
-            (pointerenter)="changeLineColorOver($event, arc)"
-            (pointerleave)="changeLineColorOut($event)"
-            (click)="onLineClick($event, arc)"
-        ></svg:path>
-        <svg:path
-            [attr.class]="'visiblePath'"
-            [attr.d]="pathForLine(arc)"
             [attr.stroke]="'black'"
             [attr.stroke-width]="width"
             [attr.fill]="'none'"
@@ -63,17 +52,17 @@ import { PetriNetManagementService } from 'src/app/services/petri-net-management
         ></svg:path>
     `,
     styles: `
-        path:hover {
-            cursor: pointer;
-        }
+        // path:hover {
+        //     cursor: pointer;
+        // }
 
-        path.active {
-            stroke: red !important;
-        }
+        // path.active {
+        //     stroke: red !important;
+        // }
 
-        path.hovered {
-            stroke: red !important;
-        }
+        // path.hovered {
+        //     stroke: red !important;
+        // }
     `,
 })
 export class DrawingArcComponent {
@@ -87,118 +76,6 @@ export class DrawingArcComponent {
 
     readonly width: number = environment.drawingElements.arcs.width;
     readonly color: string = environment.drawingElements.arcs.color;
-    rectheight = 10;
-    private timeoutId: any;
-
-    changeLineColorOver(event: Event, arc: Arc): void {
-        const pathDummy = event.target as SVGPathElement;
-        const path = pathDummy.nextElementSibling as SVGLineElement;
-        const svg: SVGSVGElement = document.getElementsByTagName(
-            'svg',
-        )[0] as SVGSVGElement;
-
-        if (this.timeoutId) {
-            clearTimeout(this.timeoutId);
-        }
-
-        if (svg && path) {
-            if (
-                !svg.classList.contains('mouseDown') &&
-                !path.classList.contains('active')
-            ) {
-                path.classList.add('hovered');
-                path.setAttribute('marker-end', 'url(#arrowhead-red)');
-            } else if (svg.classList.contains('mouseDown')) {
-                this.timeoutId = setTimeout(() => {
-                    if (this.collectSelectedElementsService.isDFGArc(arc)) {
-                        if (
-                            this.collectSelectedElementsService.isArcinSameDFG(
-                                arc,
-                            )
-                        ) {
-                            if (!path.classList.contains('active')) {
-                                path.classList.toggle('active');
-                                path.setAttribute(
-                                    'marker-end',
-                                    'url(#arrowhead-red)',
-                                );
-                            } else {
-                                path.classList.toggle('active');
-                                path.setAttribute(
-                                    'marker-end',
-                                    'url(#arrowhead)',
-                                );
-                            }
-                            this.collectSelectedElementsService.updateCollectedArcs(
-                                arc,
-                            );
-                        } else {
-                            this.feedbackService.showMessage(
-                                'Arc not in same DFG',
-                                true,
-                            );
-                        }
-                    } else {
-                        this.feedbackService.showMessage(
-                            'Arc is not a DFGArc',
-                            true,
-                        );
-                    }
-                }, 90);
-            }
-        }
-    }
-
-    changeLineColorOut(event: Event): void {
-        const svg: SVGSVGElement = document.getElementsByTagName(
-            'svg',
-        )[0] as SVGSVGElement;
-
-        if (svg && !svg.classList.contains('mouseDown')) {
-            const rect = event.target as SVGRectElement;
-            const line = rect.nextElementSibling as SVGLineElement;
-
-            if (line && !line.classList.contains('active')) {
-                line.classList.remove('hovered');
-                line.setAttribute('marker-end', 'url(#arrowhead)');
-            }
-        }
-    }
-
-    onLineClick(event: Event, arc: Arc): void {
-        const rect = event.target as SVGRectElement;
-        const line = rect.nextElementSibling as SVGLineElement;
-
-        if (this.collectSelectedElementsService.isDFGArc(arc)) {
-            if (this.collectSelectedElementsService.isArcinSameDFG(arc)) {
-                if (line) {
-                    line.classList.toggle('active');
-                }
-
-                this.collectSelectedElementsService.updateCollectedArcs(arc);
-            } else {
-                this.feedbackService.showMessage('Arc not in same DFG', true);
-            }
-        } else {
-            this.feedbackService.showMessage('Arc is not a DFGArc', true);
-        }
-    }
-
-    rectWidth(arc: Arc): number {
-        const width = Math.sqrt(
-            Math.pow(this.arc.x2 - this.arc.x1, 2) +
-                Math.pow(this.arc.y2 - this.arc.y1, 2),
-        );
-
-        return width;
-    }
-
-    get rotationAngle(): number {
-        return (
-            Math.atan2(this.arc.y2 - this.arc.y1, this.arc.x2 - this.arc.x1) *
-            (180 / Math.PI)
-        );
-    }
 
     pathForLine(arc: Arc): string {
         if (!this.pathNecessary(arc)) {
