+++ src/app/components/drawing-area/components/index.ts
@@ -1,4 +1,5 @@
 export * from '../components/drawing-activity.component';
 export * from '../components/drawing-arc.component';
+export * from '../components/drawing-boxArc.component';
 export * from '../components/drawing-place.component';
 export * from '../components/drawing-transition.component';
